requirements.txt
mss==10.1.0
numpy==2.4.2
packaging==26.0
pillow==12.1.0
pytesseract==0.3.13
